moved components into a sub-folder
